refactor: move the deploy cmd variables into its own struct
