+++ src/lib.rs
@@ -5,6 +5,7 @@
 // Please see the LICENSE file for more details.
 
 pub mod ansible;
+pub mod deploy;
 pub mod digital_ocean;
 pub mod error;
 pub mod logs;
@@ -42,7 +43,7 @@ use std::{
     os::unix::fs::PermissionsExt,
     path::{Path, PathBuf},
     process::{Command, Stdio},
-    time::{Duration, Instant},
+    time::Duration,
 };
 use tar::Archive;
 
@@ -386,155 +387,6 @@ impl TestnetDeploy {
         Ok(())
     }
 
-    pub async fn create_infra(
-        &self,
-        name: &str,
-        vm_count: u16,
-        enable_build_vm: bool,
-    ) -> Result<()> {
-        let start = Instant::now();
-        println!("Selecting {name} workspace...");
-        self.terraform_runner.workspace_select(name)?;
-        let args = vec![
-            ("node_count".to_string(), vm_count.to_string()),
-            ("use_custom_bin".to_string(), enable_build_vm.to_string()),
-        ];
-        println!("Running terraform apply...");
-        self.terraform_runner.apply(args)?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
-    pub async fn build_safe_network_binaries(
-        &self,
-        name: &str,
-        custom_branch_details: (String, String),
-    ) -> Result<()> {
-        let start = Instant::now();
-        println!("Obtaining IP address for build VM...");
-        let build_inventory = self.ansible_runner.inventory_list(
-            PathBuf::from("inventory").join(format!(".{name}_build_inventory_digital_ocean.yml")),
-        )?;
-        let build_ip = build_inventory[0].1;
-        self.ssh_client
-            .wait_for_ssh_availability(&build_ip, &self.cloud_provider.get_ssh_user())?;
-
-        println!("Running ansible against build VM...");
-        let extra_vars = self.build_binaries_extra_vars_doc(name, custom_branch_details)?;
-        self.ansible_runner.run_playbook(
-            PathBuf::from("build.yml"),
-            PathBuf::from("inventory").join(format!(".{name}_build_inventory_digital_ocean.yml")),
-            self.cloud_provider.get_ssh_user(),
-            Some(extra_vars),
-        )?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
-    pub async fn provision_genesis_node(
-        &self,
-        name: &str,
-        logstash_details: (&str, &[SocketAddr]),
-        custom_branch_details: Option<(String, String)>,
-        safenode_version: Option<String>,
-    ) -> Result<()> {
-        let start = Instant::now();
-        let genesis_inventory = self.ansible_runner.inventory_list(
-            PathBuf::from("inventory").join(format!(".{name}_genesis_inventory_digital_ocean.yml")),
-        )?;
-        let genesis_ip = genesis_inventory[0].1;
-        self.ssh_client
-            .wait_for_ssh_availability(&genesis_ip, &self.cloud_provider.get_ssh_user())?;
-        println!("Running ansible against genesis node...");
-        self.ansible_runner.run_playbook(
-            PathBuf::from("genesis_node.yml"),
-            PathBuf::from("inventory").join(format!(".{name}_genesis_inventory_digital_ocean.yml")),
-            self.cloud_provider.get_ssh_user(),
-            Some(self.build_node_extra_vars_doc(
-                name,
-                None,
-                None,
-                custom_branch_details,
-                logstash_details,
-                safenode_version,
-            )?),
-        )?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
-    pub async fn provision_faucet(
-        &self,
-        name: &str,
-        genesis_multiaddr: &str,
-        custom_branch_details: Option<(String, String)>,
-    ) -> Result<()> {
-        let start = Instant::now();
-        println!("Running ansible against genesis node to deploy faucet...");
-        self.ansible_runner.run_playbook(
-            PathBuf::from("faucet.yml"),
-            PathBuf::from("inventory").join(format!(".{name}_genesis_inventory_digital_ocean.yml")),
-            self.cloud_provider.get_ssh_user(),
-            Some(self.build_faucet_extra_vars_doc(
-                name,
-                genesis_multiaddr,
-                custom_branch_details,
-            )?),
-        )?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
-    pub async fn provision_safenode_rpc_client(
-        &self,
-        name: &str,
-        genesis_multiaddr: &str,
-        custom_branch_details: Option<(String, String)>,
-    ) -> Result<()> {
-        let start = Instant::now();
-        println!("Running ansible against genesis node to deploy safenode_rpc_client...");
-        self.ansible_runner.run_playbook(
-            PathBuf::from("safenode_rpc_client.yml"),
-            PathBuf::from("inventory").join(format!(".{name}_genesis_inventory_digital_ocean.yml")),
-            self.cloud_provider.get_ssh_user(),
-            Some(self.build_safenode_rpc_client_extra_vars_doc(
-                name,
-                genesis_multiaddr,
-                custom_branch_details,
-            )?),
-        )?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
-    pub async fn provision_remaining_nodes(
-        &self,
-        name: &str,
-        logstash_details: (&str, &[SocketAddr]),
-        genesis_multiaddr: &str,
-        node_instance_count: u16,
-        custom_branch_details: Option<(String, String)>,
-        safenode_version: Option<String>,
-    ) -> Result<()> {
-        let start = Instant::now();
-        println!("Running ansible against remaining nodes...");
-        self.ansible_runner.run_playbook(
-            PathBuf::from("nodes.yml"),
-            PathBuf::from("inventory").join(format!(".{name}_node_inventory_digital_ocean.yml")),
-            self.cloud_provider.get_ssh_user(),
-            Some(self.build_node_extra_vars_doc(
-                name,
-                Some(genesis_multiaddr.to_string()),
-                Some(node_instance_count),
-                custom_branch_details,
-                logstash_details,
-                safenode_version,
-            )?),
-        )?;
-        print_duration(start.elapsed());
-        Ok(())
-    }
-
     pub async fn get_genesis_multiaddr(&self, name: &str) -> Result<(String, IpAddr)> {
         let genesis_inventory = self.ansible_runner.inventory_list(
             PathBuf::from("inventory").join(format!(".{name}_genesis_inventory_digital_ocean.yml")),
@@ -547,94 +399,6 @@ impl TestnetDeploy {
         // The genesis_ip is obviously inside the multiaddr, but it's just being returned as a
         // separate item for convenience.
         Ok((multiaddr, genesis_ip))
-    }
-
-    pub async fn deploy(
-        &self,
-        name: &str,
-        logstash_details: (&str, &[SocketAddr]),
-        vm_count: u16,
-        node_instance_count: u16,
-        custom_branch_details: Option<(String, String)>,
-        custom_version_details: Option<(String, String)>,
-    ) -> Result<()> {
-        let safenode_version = custom_version_details.as_ref().map(|x| x.0.clone());
-        let enable_build_vm = custom_branch_details.is_some();
-
-        self.create_infra(name, vm_count, enable_build_vm)
-            .await
-            .map_err(|err| {
-                println!("Failed to create infra {err:?}");
-                err
-            })?;
-        if let Some(custom_branch) = &custom_branch_details {
-            self.build_safe_network_binaries(name, custom_branch.clone())
-                .await
-                .map_err(|err| {
-                    println!("Failed to build safe network binaries {err:?}");
-                    err
-                })?;
-        }
-
-        self.provision_genesis_node(
-            name,
-            logstash_details,
-            custom_branch_details.clone(),
-            safenode_version.clone(),
-        )
-        .await
-        .map_err(|err| {
-            println!("Failed to provision genesis node {err:?}");
-            err
-        })?;
-
-        let (multiaddr, _) = self.get_genesis_multiaddr(name).await.map_err(|err| {
-            println!("Failed to get genesis multiaddr {err:?}");
-            err
-        })?;
-        println!("Obtained multiaddr for genesis node: {multiaddr}");
-
-        self.provision_remaining_nodes(
-            name,
-            logstash_details,
-            &multiaddr,
-            node_instance_count,
-            custom_branch_details.clone(),
-            safenode_version,
-        )
-        .await
-        .map_err(|err| {
-            println!("Failed to provision remaining nodes {err:?}");
-            err
-        })?;
-
-        self.provision_faucet(name, &multiaddr, custom_branch_details.clone())
-            .await
-            .map_err(|err| {
-                println!("Failed to provision faucet {err:?}");
-                err
-            })?;
-
-        self.provision_safenode_rpc_client(name, &multiaddr, custom_branch_details.clone())
-            .await
-            .map_err(|err| {
-                println!("Failed to provision safenode rpc client {err:?}");
-                err
-            })?;
-
-        self.list_inventory(
-            name,
-            true,
-            custom_branch_details,
-            custom_version_details,
-            Some(node_instance_count),
-        )
-        .await
-        .map_err(|err| {
-            println!("Failed to list inventory {err:?}");
-            err
-        })?;
-        Ok(())
     }
 
     /// Print the inventory for the deployment.
@@ -781,175 +545,6 @@ impl TestnetDeploy {
                 "node".to_string(),
             ],
         )
-    }
-
-    ///
-    /// Private Helpers
-    ///
-    fn build_node_extra_vars_doc(
-        &self,
-        name: &str,
-        genesis_multiaddr: Option<String>,
-        node_instance_count: Option<u16>,
-        custom_branch_details: Option<(String, String)>,
-        logstash_details: (&str, &[SocketAddr]),
-        safenode_version: Option<String>,
-    ) -> Result<String> {
-        let mut extra_vars = String::new();
-        extra_vars.push_str("{ ");
-        Self::add_value(
-            &mut extra_vars,
-            "provider",
-            &self.cloud_provider.to_string(),
-        );
-        Self::add_value(&mut extra_vars, "testnet_name", name);
-        if genesis_multiaddr.is_some() {
-            Self::add_value(
-                &mut extra_vars,
-                "genesis_multiaddr",
-                &genesis_multiaddr.ok_or_else(|| Error::GenesisMultiAddrNotSupplied)?,
-            );
-        }
-        if node_instance_count.is_some() {
-            Self::add_value(
-                &mut extra_vars,
-                "node_instance_count",
-                &node_instance_count.unwrap_or(20).to_string(),
-            );
-        }
-        if let Some((repo_owner, branch)) = custom_branch_details {
-            Self::add_value(
-                &mut extra_vars,
-                "node_archive_url",
-                &format!(
-                    "https://sn-node.s3.eu-west-2.amazonaws.com/{}/{}/safenode-{}-x86_64-unknown-linux-musl.tar.gz",
-                    repo_owner,
-                    branch,
-                    name),
-            );
-        }
-        if let Some(version) = safenode_version {
-            Self::add_value(
-                &mut extra_vars,
-                "node_archive_url",
-                &format!(
-                    "https://github.com/maidsafe/safe_network/releases/download/sn_node-v{version}/safenode-{version}-x86_64-unknown-linux-musl.tar.gz",
-                ),
-            );
-        }
-
-        let (logstash_stack_name, logstash_hosts) = logstash_details;
-        Self::add_value(&mut extra_vars, "logstash_stack_name", logstash_stack_name);
-        extra_vars.push_str("\"logstash_hosts\": [");
-        for host in logstash_hosts.iter() {
-            extra_vars.push_str(&format!("\"{}\", ", host));
-        }
-        let mut extra_vars = extra_vars.strip_suffix(", ").unwrap().to_string();
-        extra_vars.push_str("] }");
-        Ok(extra_vars)
-    }
-
-    fn build_faucet_extra_vars_doc(
-        &self,
-        name: &str,
-        genesis_multiaddr: &str,
-        custom_branch_details: Option<(String, String)>,
-    ) -> Result<String> {
-        let mut extra_vars = String::new();
-        extra_vars.push_str("{ ");
-        Self::add_value(
-            &mut extra_vars,
-            "provider",
-            &self.cloud_provider.to_string(),
-        );
-        Self::add_value(&mut extra_vars, "testnet_name", name);
-        Self::add_value(&mut extra_vars, "genesis_multiaddr", genesis_multiaddr);
-        if let Some((repo_owner, branch)) = custom_branch_details {
-            Self::add_value(&mut extra_vars, "branch", &branch);
-            Self::add_value(&mut extra_vars, "org", &repo_owner);
-            Self::add_value(
-                &mut extra_vars,
-                "faucet_archive_url",
-                &format!(
-                    "https://sn-node.s3.eu-west-2.amazonaws.com/{}/{}/faucet-{}-x86_64-unknown-linux-musl.tar.gz",
-                    repo_owner,
-                    branch,
-                    name),
-            );
-        } else {
-            Self::add_value(
-                &mut extra_vars,
-                "faucet_archive_url",
-                "https://sn-faucet.s3.eu-west-2.amazonaws.com/faucet-latest-x86_64-unknown-linux-musl.tar.gz",
-            );
-        }
-
-        let mut extra_vars = extra_vars.strip_suffix(", ").unwrap().to_string();
-        extra_vars.push_str(" }");
-        Ok(extra_vars)
-    }
-
-    fn build_safenode_rpc_client_extra_vars_doc(
-        &self,
-        name: &str,
-        genesis_multiaddr: &str,
-        custom_branch_details: Option<(String, String)>,
-    ) -> Result<String> {
-        let mut extra_vars = String::new();
-        extra_vars.push_str("{ ");
-        Self::add_value(
-            &mut extra_vars,
-            "provider",
-            &self.cloud_provider.to_string(),
-        );
-        Self::add_value(&mut extra_vars, "testnet_name", name);
-        Self::add_value(&mut extra_vars, "genesis_multiaddr", genesis_multiaddr);
-        if let Some((repo_owner, branch)) = custom_branch_details {
-            Self::add_value(&mut extra_vars, "branch", &branch);
-            Self::add_value(&mut extra_vars, "org", &repo_owner);
-            Self::add_value(
-                &mut extra_vars,
-                "safenode_rpc_client_archive_url",
-                &format!(
-                    "https://sn-node.s3.eu-west-2.amazonaws.com/{}/{}/safenode_rpc_client-{}-x86_64-unknown-linux-musl.tar.gz",
-                    repo_owner,
-                    branch,
-                    name),
-            );
-        } else {
-            Self::add_value(
-                &mut extra_vars,
-                "safenode_rpc_client_archive_url",
-                "https://sn-node-rpc-client.s3.eu-west-2.amazonaws.com/safenode_rpc_client-latest-x86_64-unknown-linux-musl.tar.gz",
-            );
-        }
-
-        let mut extra_vars = extra_vars.strip_suffix(", ").unwrap().to_string();
-        extra_vars.push_str(" }");
-        Ok(extra_vars)
-    }
-
-    fn build_binaries_extra_vars_doc(
-        &self,
-        name: &str,
-        (repo_owner, branch): (String, String),
-    ) -> Result<String> {
-        let mut extra_vars = String::new();
-        extra_vars.push_str("{ ");
-
-        Self::add_value(&mut extra_vars, "custom_bin", "true");
-        Self::add_value(&mut extra_vars, "branch", &branch);
-        Self::add_value(&mut extra_vars, "org", &repo_owner);
-        Self::add_value(&mut extra_vars, "testnet_name", name);
-
-        let mut extra_vars = extra_vars.strip_suffix(", ").unwrap().to_string();
-        extra_vars.push_str(" }");
-
-        Ok(extra_vars)
-    }
-
-    fn add_value(document: &mut String, name: &str, value: &str) {
-        document.push_str(&format!("\"{name}\": \"{value}\", "))
     }
 }
